migrate types for usercontroller/apikey to types folder
